Map changing behavior changes
Updated the code for changing maps.  We’re still not at our ideal way
of changing maps, but we’re getting closer.

I don’t really like the fact that sTileIndex is declared in main(), but
I’m sticking with it until I can figure out a way to load the ncurses
extended character set sooner…