Minor update to close out current branch
This is a minor update to push the remaining changes for the branch to
master.  I’m starting to work on things outside of what this branch was
originally intended for, the stair logic is pretty good at this point,
although I would like to revisit it again later after I’ve fixed the
maps.